fix(search): build errors
